+++ src/main/java/it/gov/pagopa/fdr/service/middleware/validator/SemanticValidator.java
@@ -5,16 +5,16 @@ import it.gov.pagopa.fdr.controller.model.payment.request.AddPaymentRequest;
 import it.gov.pagopa.fdr.controller.model.payment.request.DeletePaymentRequest;
 import it.gov.pagopa.fdr.exception.AppException;
 import it.gov.pagopa.fdr.repository.entity.flow.FdrFlowEntity;
-import it.gov.pagopa.fdr.service.middleware.validator.clause.AddPaymentRequestValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.BrokerPspValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.ChannelValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.ComputedPaymentsValidator;
-import it.gov.pagopa.fdr.service.middleware.validator.clause.CreateFlowRequestValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.CreditorInstitutionValidator;
-import it.gov.pagopa.fdr.service.middleware.validator.clause.DeletePaymentRequestValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.FlowNameFormatValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.PspValidator;
 import it.gov.pagopa.fdr.service.middleware.validator.clause.PublishableStatusValidator;
+import it.gov.pagopa.fdr.service.middleware.validator.clause.request.AddPaymentRequestValidator;
+import it.gov.pagopa.fdr.service.middleware.validator.clause.request.CreateFlowRequestValidator;
+import it.gov.pagopa.fdr.service.middleware.validator.clause.request.DeletePaymentRequestValidator;
 import it.gov.pagopa.fdr.service.model.FindFlowsByFiltersArgs;
 import it.gov.pagopa.fdr.util.StringUtil;
 import it.gov.pagopa.fdr.util.validator.ValidationArgs;
