+++ src/main/java/it/gov/pagopa/fdr/service/middleware/validator/clause/request/DeletePaymentRequestValidator.java
@@ -1,27 +1,26 @@
-package it.gov.pagopa.fdr.service.middleware.validator.clause;
+package it.gov.pagopa.fdr.service.middleware.validator.clause.request;
 
-import it.gov.pagopa.fdr.controller.model.payment.Payment;
-import it.gov.pagopa.fdr.controller.model.payment.request.AddPaymentRequest;
+import it.gov.pagopa.fdr.controller.model.payment.request.DeletePaymentRequest;
 import it.gov.pagopa.fdr.exception.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdr.util.validator.ValidationArgs;
 import it.gov.pagopa.fdr.util.validator.ValidationResult;
 import it.gov.pagopa.fdr.util.validator.ValidationStep;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
-public class AddPaymentRequestValidator extends ValidationStep {
+public class DeletePaymentRequestValidator extends ValidationStep {
 
   @Override
   public ValidationResult validate(ValidationArgs args) {
 
     String flowName = args.getArgument("flowName", String.class);
-    AddPaymentRequest request = args.getArgument("request", AddPaymentRequest.class);
+    DeletePaymentRequest request = args.getArgument("request", DeletePaymentRequest.class);
 
     // check if there are some duplicated indexes in request
-    List<Payment> payments = request.getPayments();
-    Set<Long> uniqueIndexes = payments.stream().map(Payment::getIndex).collect(Collectors.toSet());
-    if (payments.size() != uniqueIndexes.size()) {
+    List<Long> indexes = request.getIndexList();
+    Set<Long> uniqueIndexes = new HashSet<>(indexes);
+    if (indexes.size() != uniqueIndexes.size()) {
       return ValidationResult.asInvalid(
           AppErrorCodeMessageEnum.REPORTING_FLOW_PAYMENT_SAME_INDEX_IN_SAME_REQUEST, flowName);
     }
